:art: design - 페이지내 기본 컬러 적용
- base color : #25b6b1
