+++ src/features/profile/index.tsx
@@ -25,8 +25,8 @@ const Profile = () => {
         className="max-w-4xl mx-auto flex flex-col items-center justify-center"
         titleClassName="hidden"
         fullHeight={true}>
-        <article className="mt-40">
-          <h3 className="p-10 text-center text-3xl font-semibold">
+        <article>
+          <h3 className="p-10 text-primary text-center text-3xl font-semibold">
             ECO FRIENDLY
           </h3>
           <h4 className="mt-10 text-center text-2xl font-semibold">
@@ -74,10 +74,10 @@ const Profile = () => {
       {/* Contact Us */}
       <SectionLayout
         title="CONTACT US"
-        className="max-w-4xl mx-auto flex flex-col items-center justify-center"
+        className="max-w-6xl mx-auto flex flex-col items-center justify-center"
         titleClassName="text-6xl font-normal"
         fullHeight={true}>
-        <p className="max-w-2xl mb-5 text-center text-lg text-gray-700 leading-relaxed">
+        <p className="max-w-2xl mb-16 text-center text-lg text-gray-700 leading-relaxed">
           궁금한 점이 있으신가요? 언제든지 문의해 주세요.
           <br />
           비젼라이프홀딩스는 고객과의 소통을 중요하게 생각하며,
@@ -90,25 +90,25 @@ const Profile = () => {
             <ListCard
               title="전화번호"
               type="_base"
-              icon={<FaPhoneAlt className="w-14 h-14 p-3 border rounded-full"/>}
+              icon={<FaPhoneAlt className="w-14 h-14 p-3 bg-primary text-white rounded-full"/>}
               description="02-555-1202"
               size="w-full h-full"/>
             <ListCard
               title="팩스 (FAX)"
               type="_base"
-              icon={<LiaFaxSolid className="w-14 h-14 p-3 border rounded-full"/>}
+              icon={<LiaFaxSolid className="w-14 h-14 p-3 bg-primary text-white rounded-full"/>}
               description="02-555-5801"
               size="w-full h-full"/>
             <ListCard
               title="이메일"
               type="_base"
-              icon={<MdOutlineMail className="w-14 h-14 p-3 border rounded-full"/>}
+              icon={<MdOutlineMail className="w-14 h-14 p-3 bg-primary text-white rounded-full"/>}
               description="visionlife2018@naver.com"
               size="w-full h-full"/>
             <ListCard
               title="주소"
               type="_base"
-              icon={<HiOutlineLocationMarker className="w-16 h-14 p-3 border rounded-full"/>}
+              icon={<HiOutlineLocationMarker className="w-16 h-14 p-3 bg-primary text-white rounded-full"/>}
               description="서울 서초구 반포대로 21길 23 세진빌딩 3층"
               size="w-full h-full"/>
           </ul>
