+++ src/index.css
@@ -1,4 +1,7 @@
 @import 'tailwindcss';
+@theme {
+  --color-primary: #25b6b1
+}
 
 body {
   overflow-x: hidden;
